revert my data type mistake

+++ db/migrations/00010_create_eth_state_accouts_table.sql
@@ -2,8 +2,8 @@
 CREATE TABLE eth.state_accounts (
     header_id             VARCHAR(66) NOT NULL,
     state_path            BYTEA NOT NULL,
-    balance               BIGINT NOT NULL,
-    nonce                 INTEGER NOT NULL,
+    balance               NUMERIC NOT NULL,
+    nonce                 BIGINT NOT NULL,
     code_hash             BYTEA NOT NULL,
     storage_root          VARCHAR(66) NOT NULL,
     FOREIGN KEY (header_id, state_path) REFERENCES eth.state_cids (header_id, state_path) ON DELETE CASCADE DEFERRABLE INITIALLY DEFERRED,
